Move up Pid command
